refactor(review): review,reply 서비스 분리, 불필요 부분 삭제
